validation tests: fix type mismatch
The asm uses 'x4', which is a 64-bit register, but it is given the
variable 'n', which is an int (only 32 bits). Due to this, only the
lower 32 bits of the register are guaranteed to be valid.

These changes were tested on the ARM Cortex-A72 architecture.

+++ src/validation_tests/load_store_testcode.c
@@ -1,7 +1,7 @@
 #include "testcode.h"
 
 /* Execute n stores */
-int execute_stores(int n) {
+int execute_stores(long long n) {
 
 #if defined(__aarch64__)
 
@@ -30,7 +30,7 @@ int execute_stores(int n) {
 }
 
 /* Execute n loads */
-int execute_loads(int n) {
+int execute_loads(long long n) {
 
 #if defined(__aarch64__)
